+++ lib/src/constants/regex_constants.dart
@@ -8,8 +8,6 @@ class RegexConstants {
     return _instance!;
   }
 
-  final String emailRegex =
-      r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+";
-  final String passwordRegex =
-      r'^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[!@#\$&*~]).{8,}$';
+  final String emailRegex = r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$';
+  final String passwordRegex = r'^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[!@#\$&*~]).{8,}$';
 }